Separate row into container and container components
