+++ lib/react/containers/row.js
@@ -4,7 +4,7 @@ import { connect } from "react-redux";
 import type { Dispatch } from "redux";
 import type { MapDispatchToProps } from "react-redux";
 import * as A from "../../action-creators";
-import Row from "../row";
+import Row from "../presentationals/row";
 
 const mapDispatchToProps: MapDispatchToProps<Action, *, *> = (
   dispatch: Dispatch<Action>
