+++ lib/react/panel.js
@@ -4,6 +4,7 @@ import classNames from "classnames";
 import React from "react";
 import ReactDOM from "react-dom";
 import { Observable } from "rxjs";
+import Cell from "./containers/cell";
 import Loading from "./presentationals/loading";
 import Row from "./containers/row";
 import ResizeHandle from "./containers/resize-handle";
@@ -62,7 +63,9 @@ export default class Panel extends React.Component<Props> {
                 })}
               >
                 {this.props.visibleRows.map(row => (
-                  <Row key={row.id} row={row} />
+                  <Row key={row.id} row={row}>
+                    {row.cells.map((cell, i) => <Cell key={i} cell={cell} />)}
+                  </Row>
                 ))}
               </tbody>
             </table>
