+++ lib/react/presentationals/row.js
@@ -1,10 +1,10 @@
 /* @flow */
 
 import classNames from "classnames";
-import React from "react";
-import Cell from "./containers/cell";
+import * as React from "react";
 
 type Props = {
+  children?: React.Node,
   row: NoteRow,
   onClick: (row: NoteRow) => any
 };
@@ -19,7 +19,7 @@ export default class Row extends React.Component<Props> {
         }}
         className={classNames({ "is-selected": row.selected })}
       >
-        {row.cells.map((cell, i) => <Cell key={i} cell={cell} />)}
+        {this.props.children}
       </tr>
     );
   }
